Rename 'components/PositiveRateMixedChart.vue' to 'components/index/CardsMonitoring/PositiveRate/Chart.vue'
